windows_terminal_settings.json: Replace "Ubuntu-20.04" with "Ubuntu".
"Ubuntu" from the Microsoft Store will always be the latest LTS version (see https://www.howtogeek.com/fyi/there-are-now-three-versions-of-ubuntu-in-the-microsoft-store-heres-why/).

Updating windows_terminal_settings.json was not as simple as:
    -                "name": "Ubuntu-20.04",
    +                "name": "Ubuntu",
                     "guid": "{07b52e3e-de2c-5db4-bd2d-ba144ed6c273}",
                     "source": "Windows.Terminal.Wsl",
    -                "startingDirectory": "\\\\wsl$\\Ubuntu-20.04\\home\\shane"
    +                "startingDirectory": "\\\\wsl$\\Ubuntu\\home\\shane"
That would result in the "Ubuntu" profile being ignored. The Ubuntu profile was ignored because of the guid; Profiles that specify a "source" (such as "Windows.Terminal.Wsl") are "dynamic profiles" and the guid of such a profile depends on the distro's name (07b52e3e-de2c-5db4-bd2d-ba144ed6c273 is specific to Ubuntu-20.04, see: https://github.com/microsoft/terminal/issues/2931#issuecomment-535922476).

An alternative to using a dynamic profile:
-                "source": "Windows.Terminal.Wsl",
+                "icon": "ms-appx:///ProfileIcons/{9acb9455-ca41-5af7-950f-6bca1bc9722f}.png",
+                "colorScheme": "Campbell",
+                "commandline": "wsl.exe -d Ubuntu",
(and update the profile's guid).