Highlight hyperlinks in Vim help files.
For a long time I've "cleared" highlighting of things which are linked to the Identifier highlight group. This works well for me with syntax highlighting of source files, but helpHyperTextJump was also linked to the Identifier highlight group which Clear_unwanted_syntax_highlighting() cleared; now Clear_unwanted_syntax_highlighting() "restores" the highlighting for helpHyperTextJump.