.vimrc: Delete "inoremap <esc> <nop>".
I have been using the "jk" mapping to exit insert mode, but that's not a reason to disable the function of the Esc key.

(With my vim-code_block_markers plugin running inserting text that ends with 'j' (which almost never happens) requires that one wait a moment before typing 'jk' otherwise the plugin's mapping for 'jj' will be used and the j last inserted will be removed. Esc could now be used after inserting a 'j'. It'd be nice rather than map j followed by k to Esc to map j and k pressed together (making a "chord") to Esc.)