install script: Add hgd and gd Bash functions to show Mercurial and Git diffs, respectively, using Vim as the pager.
I use Vim as a pager because of its syntax highlighting. I've created a Vim color scheme that is really good for viewing diffs in monochrome (removed lines are displayed with "strikethrough" style) which helps a lot when using my monochrome E Ink monitor. Note: Previously with a diff that was short enough to fit entirely on the current terminal the pager would exit immediately however Vim doesn't do that - see https://stackoverflow.com/questions/19006704/how-to-exit-automatically-at-end-of-file-on-vim for a workaround.

vimrc: Remove normal mode mapping definition for q<space> so that pressing 'q' in vimless will cause it to exit immediately without it waiting to see if the user would then type a space. With Vim's timeout and timeoutlen options set to their default values Vim would wait a second before exiting if no key was pressed after 'q'. I hadn't been using the q<space> mapping anyway.