windows_terminal_settings.json: Disable multiLinePasteWarning.
multiLinePasteWarning was enabled by default. To paste text containing newline characters at the Ubuntu shell prompt it was necessary to confirm that it was ok to proceeded. The confirmation window showed a preview of the clipboard's contents - I will now have to be extra careful with what I paste. I noticed that the warning window wasn't displayed when pasting multiple lines into Vim.