windows_terminal_settings.json: Reenable multiLinePasteWarning and disable largePasteWarning.
Reenable multiLinePasteWarning (now explicitly enabled): Text copied to the Windows clipboard from a website could contain something unexpected. When pasting into a shell with multiLinePasteWarning enabled a preview of the text to be pasted will be shown in a confirmation window. No confirmation window is shown when pasting into Vim.

With largePasteWarning enabled (the default) a warning window is displayed to ask for confirmation when pasting >5 KiB of text. The option is now disabled.