Change color used for username and hostname at the start of the bash prompt.
PS1 set in ~/.bash is overridden by bashrc.

A green with a bit of blue is now used (color #29 from the 256-color palette). Previously green from the 8-color terminal palette was used. With the 8-color palette the exact color used could depend on terminal settings while colors in the 256-color palette (excluding the first 16) should be fixed. (With Windows Terminal the colors used for the 8-color palette depend on the color scheme selected and also on the intenseTextStyle setting - "intense text" with intenseTextStyle="all" or "bright" will use a different entry from the color palette than when intenseTextStyle="none" or "bold.)