gd, gsh, hgd and hge functions: stderr no longer redirected to stdout.
This reverts the change made in commit c43a006d.

Now, for example, running hgd from a directory that's not part of a Mercurial repository results in "abort: no repository found..." being output to stderr and vimless is not run. Previously vimless was run and it showed the error message that had been redirected from stderr.