vim/colors/sgh.vim: Define helpExample highlighting group.
By default helpExample is linked to Comment. Linking helpExample to Comment doesn't make sense; example code in the Vim help isn't a comment. If someone were to set the color of comments (e.g. to either stand-out or to not stand-out) they'd most likely not expect or want the appearance of example code in the Vim help to also change.