bashrc: hge: shorten header text.
The header is now 2 lines; previously 6 lines were used.

I tried leaving out "# HG changeset patch" from the start of the first line but then Vim didn't set filetype to diff and consequently syntax highlighting was not as expected.

The branch name is only included if it's not "default" (as for the standard "hg export" template). The author is only included if I'm not the author.