gitignore + hgignore: Don't match a .swg file (SWIG interface file) as a VIM swap file.
ag ("The Silver Searcher") uses gitignore and hgignore files to get a list of files to ignore; I had ag miss matches I'd expected to see in a .swg file. This commit fixes that problem.

https://swig-experimental.readthedocs.io/en/latest/Manual/SWIG.html#input-format mentions that SWIG interface files may have a ".i" or ".swg" filename extension.

Note: Now if there are more than 3 Vim swap files for a particular file then the 4th, 5th, etc. swap files won't be ignored - I don't expect that to be a problem though.