install: Don't abort if ~/.vimrc or ~/.dircolors already exist.
This change makes it possible for the install script to run all the way through if its rerun, e.g. after it's updated. Previously the script would have failed with:
    ln: failed to create symbolic link '/home/shane/.vimrc': File exists
Now,
    ln: replace '/home/shane/.vimrc'?
will be asked. (A similar question for .dircolors will also be asked.)

Ideally the questions wouldn't be asked if .vimrc and .dircolors won't be changed, i.e. they're already symlinks pointing to ./vimrc and ./dircolors respectively.